server/ai_agent/requirements.txt
livekit-agents>=0.6.0
livekit-plugins-deepgram
livekit-plugins-openai
python-dotenv
